cliff(...) gets its own file =D
